refactor: add js extension to relative imports
