+++ src/package/components/Badge/__tests__/Badge.test.tsx
@@ -1,7 +1,7 @@
 import { screen } from "@testing-library/dom";
 import "@testing-library/jest-dom/extend-expect";
 import { render } from "@testing-library/react";
-import { Badge } from "../Badge";
+import { Badge } from "../Badge.js";
 
 describe("<Badge>", () => {
   test("Renders Badge tag with children text", () => {
