+++ src/package/components/Button/__tests__/Button.test.tsx
@@ -1,7 +1,7 @@
 import { screen } from "@testing-library/dom";
 import "@testing-library/jest-dom/extend-expect";
 import { render } from "@testing-library/react";
-import { Button } from "../Button";
+import { Button } from "../Button.js";
 
 describe("<Button>", () => {
   test("Renders button tag with children text", () => {
