+++ src/package/components/Card/Card.tsx
@@ -12,8 +12,8 @@ import {
   TypographyProps as SSTypographyProps,
 } from "styled-system";
 
-import { Color } from "../../foundation/Color/Color";
-import { Shadow } from "../../foundation/Shadow/Shadow";
+import { Color } from "../../foundation/Color/Color.js";
+import { Shadow } from "../../foundation/Shadow/Shadow.js";
 
 export interface CardProps
   extends FlexboxProps,
