--- src/package/components/Search/CustomComponents/index.ts
@@ -1,4 +1,0 @@
-export { ClearIndicator } from "./ClearIndicator";
-export { Control } from "./Control";
-export { IndicatorsContainer } from "./IndicatorsContainer";
-export { NoOptionsMessage } from "./NoOptionsMessage";
