+++ src/package/components/Search/stories/examples/EgFiltering.tsx
@@ -1,6 +1,6 @@
-import { showToast } from "../../../Toast/Toast";
-import { Search } from "../../Search";
-import { SearchOptionType, SearchValueType } from "../../types";
+import { showToast } from "../../../Toast/Toast.js";
+import { Search } from "../../Search.js";
+import { SearchOptionType, SearchValueType } from "../../types.js";
 
 export function EgFiltering() {
   return (
