+++ src/package/components/Search/stories/examples/EgSingleSelect.tsx
@@ -1,6 +1,6 @@
-import { showToast } from "../../../Toast/Toast";
-import { Search } from "../../Search";
-import { SearchOptionType, SearchValueType } from "../../types";
+import { showToast } from "../../../Toast/Toast.js";
+import { Search } from "../../Search.js";
+import { SearchOptionType, SearchValueType } from "../../types.js";
 
 export function EgSingleSelect() {
   return (
