+++ src/package/components/Slider/Slider.tsx
@@ -1,8 +1,8 @@
 import RCSlider, { Handle, SliderProps as RCSliderProps } from "rc-slider";
 import styled from "styled-components";
 
-import { Box } from "../../foundation/Box/Box";
-import { Icon } from "../../foundation/Icon/Icon";
+import { Box } from "../../foundation/Box/Box.js";
+import { Icon } from "../../foundation/Icon/Icon.js";
 
 export interface SliderProps extends RCSliderProps {
   width?: string;
