+++ src/package/foundation/MediaQuery/stories/MediaQuery.stories.mdx
@@ -1,6 +1,9 @@
 import { ArgsTable, Canvas, Meta, Story } from "@storybook/addon-docs";
+
 import { MediaQuery } from "../MediaQuery";
-import { EgMobileFirst, EgDesktopFirst, EgTouch } from "./examples";
+import { EgMobileFirst } from "./examples/EgMobileFirst";
+import { EgDesktopFirst } from "./examples/EgDesktopFirst";
+import { EgTouch } from "./examples/EgTouch";
 
 <Meta title="Foundation/MediaQuery" component={MediaQuery} />
