--- src/package/foundation/MediaQuery/stories/examples/index.ts
@@ -1,3 +1,0 @@
-export { EgDesktopFirst } from "./EgDesktopFirst";
-export { EgMobileFirst } from "./EgMobileFirst";
-export { EgTouch } from "./EgTouch";
