+++ src/package/foundation/Typography/Typography.tsx
@@ -1,10 +1,11 @@
+// @ts-ignore
 import styled from "styled-components";
 
-import { applyTypographyVariant } from "./styles";
-import { TextTag, TypographyType } from "./types";
+import { applyTypographyVariant } from "./styles.js";
+import { TextTag, TypographyType } from "./types.js";
 
-import type { Color } from "../Color/Color";
-import type { SxProps } from "../sx";
+import type { Color } from "../Color/Color.js";
+import type { SxProps } from "../sx.js";
 import type { ReactNode } from "react";
 
 export interface TypographyProps {
@@ -15,8 +16,12 @@ export interface TypographyProps {
   sx?: SxProps;
 }
 
+// eslint-disable-next-line @typescript-eslint/ban-ts-comment
+// @ts-ignore
 const StyledDiv = styled.div<TypographyProps>`
-  ${(props) => props.variant && applyTypographyVariant(props.variant)};
+  ${(props: { variant?: TypographyType }) =>
+    // eslint-disable-next-line implicit-arrow-linebreak
+    props.variant && applyTypographyVariant(props.variant)};
 `;
 
 export function Typography({
