--- src/package/foundation/Typography/responsive/index.ts
@@ -1,1 +1,0 @@
-export { letterSpacing } from "./letterSpacing";
