+++ src/package/index.ts
@@ -1,45 +1,45 @@
-export { Anchor } from "./components/Anchor/Anchor";
-export type { AnchorProps } from "./components/Anchor/Anchor";
+export { Anchor } from "./components/Anchor/Anchor.js";
+export type { AnchorProps } from "./components/Anchor/Anchor.js";
 
-export { Badge } from "./components/Badge/Badge";
-export type { BadgeProps } from "./components/Badge/Badge";
+export { Badge } from "./components/Badge/Badge.js";
+export type { BadgeProps } from "./components/Badge/Badge.js";
 
-export { Button } from "./components/Button/Button";
-export type { ButtonProps } from "./components/Button/Button";
+export { Button } from "./components/Button/Button.js";
+export type { ButtonProps } from "./components/Button/Button.js";
 
-export { Card } from "./components/Card/Card";
-export type { CardProps } from "./components/Card/Card";
+export { Card } from "./components/Card/Card.js";
+export type { CardProps } from "./components/Card/Card.js";
 
-export { Search } from "./components/Search/Search";
-export type { SearchProps } from "./components/Search/Search";
+export { Search } from "./components/Search/Search.js";
+export type { SearchProps } from "./components/Search/Search.js";
 export type {
   SearchOptionType,
   SearchValueType,
-} from "./components/Search/types";
+} from "./components/Search/types.js";
 
-export { Toast, showToast } from "./components/Toast/Toast";
+export { Toast, showToast } from "./components/Toast/Toast.js";
 
-export { Slider } from "./components/Slider/Slider";
-export type { SliderProps } from "./components/Slider/Slider";
+export { Slider } from "./components/Slider/Slider.js";
+export type { SliderProps } from "./components/Slider/Slider.js";
 
-export { Table } from "./components/Table/Table";
-export type { TableProps } from "./components/Table/Table";
+export { Table } from "./components/Table/Table.js";
+export type { TableProps } from "./components/Table/Table.js";
 
-export { Tooltip } from "./components/Tooltip/Tooltip";
-export type { TooltipProps } from "./components/Tooltip/Tooltip";
+export { Tooltip } from "./components/Tooltip/Tooltip.js";
+export type { TooltipProps } from "./components/Tooltip/Tooltip.js";
 
 /* **********************
  * Foundation components
  *********************** */
 
-export { Box } from "./foundation/Box/Box";
-export type { BoxProps } from "./foundation/Box/Box";
+export { Box } from "./foundation/Box/Box.js";
+export type { BoxProps } from "./foundation/Box/Box.js";
 
-export { Color } from "./foundation/Color/Color";
-export { Icon } from "./foundation/Icon/Icon";
-export { MediaQuery } from "./foundation/MediaQuery/MediaQuery";
-export { Shadow } from "./foundation/Shadow/Shadow";
-export { Theme } from "./foundation/Theme/Theme";
+export { Color } from "./foundation/Color/Color.js";
+export { Icon } from "./foundation/Icon/Icon.js";
+export { MediaQuery } from "./foundation/MediaQuery/MediaQuery.js";
+export { Shadow } from "./foundation/Shadow/Shadow.js";
+export { Theme } from "./foundation/Theme/Theme.js";
 
-export { Typography } from "./foundation/Typography/Typography";
-export type { TypographyProps } from "./foundation/Typography/Typography";
+export { Typography } from "./foundation/Typography/Typography.js";
+export type { TypographyProps } from "./foundation/Typography/Typography.js";
